+++ src/components/Pages/TermsOfUse/Current.vue
@@ -1292,7 +1292,7 @@
 </template>
 
 <script>
-import TermsOfUseNavigationPanel from './NavigationPanel.vue'
+import TermsOfUseNavigationPanel from './TermsOfUseNavigationPanel.vue'
 
 export default {
   name: 'TermsOfUse',
